nhat update 12 02

+++ app/Http/Controllers/ActorController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
 use App\Models\Actor;
+use App\Models\User;
 use Illuminate\Support\Facades\DB;
 
 class ActorController extends Controller
@@ -12,7 +13,7 @@ class ActorController extends Controller
 
     public function create(Request $request)
     {
-        $this->authorize('create', Actor::class);
+        $this->authorize('onlyAdmin', Actor::class);
 
         $validator = Validator::make($request->all(), [
             'name' => 'required|string',
@@ -50,10 +51,17 @@ class ActorController extends Controller
 
     public function update(Request $request, Actor $actor)
     {
+        $this->authorize('onlyAdmin');
+
         $validator = Validator::make($request->all(), [
             'name' => 'string',
-            'popularity' => 'numeric',
+            'birthday' => 'date',
+            'gender' => 'in:female,male',
+            'place_of_birth' => 'string',
             'profile_path' => 'string',
+            'biography' => 'string',
+            'imdb_id' => 'string',
+            'popularity' => 'numeric',
         ]);
 
         if ($validator->fails()) {
@@ -63,7 +71,16 @@ class ActorController extends Controller
             ]);
         }
 
-        $actor->update($request->all());
+        $actor->update([
+            'name' => $request->input('name'),
+            'birthday' => $request->input('birthday'),
+            'gender' => $request->input('gender'),
+            'place_of_birth' => $request->input('place_of_birth'),
+            'profile_path' => $request->input('profile_path'),
+            'biography' => $request->input('biography'),
+            'imdb_id' => $request->input('imdb_id'),
+            'popularity' => $request->input('popularity'),
+        ]);
 
         return response()->json([
             'success' => true,
@@ -73,6 +90,9 @@ class ActorController extends Controller
 
     public function destroy(Request $request, Actor $actor)
     {
+
+        $this->authorize('onlyAdmin');
+
         $actor->delete();
 
         return response()->json([
